Remove the tls secret name from prod
This is because the tls cert secret is no longer available (doesn't exist)

We are thankfully being provided a tls cert by cloud platform. They
use a wild card to match many domains in it's cluster that don't
have a tls cert and providees a default one.

Add a todo to look into usng a custom tls cert