PIC-4424 Consume Large SNS messages via S3 (#704)
* PIC-4424 Configure S3AsyncClient

Configure the s3AsyncClient which will be used to get large sns
messages from S3

* Add S3 service to get object from S3

Use an s3 service to get large sns messages stored in s3.

Update the hearing extractor to use the json object in s3
as the sns message body for parsing.

* Add tests

* Refactor code

* PIC-4424 Remove feature flag

Remove passHearingIdToCourtCaseService as it's now enabled in all
environments

* PIC-4424 Reuse objectmapper bean

* PIC-4424 Update test to mock objectmapping

Update test to mock the objectMapper bean which reads and parses
the S3 pointer stored in the SNS message

* PIC-4424 Remove unused builder annotation

* PIC-4424 Add integration test

* PIC-4424 Use custom eventype for processing

Use the custom messageAtrribute eventType to decide how
 to process a hearing payload

 Add tests

* PIC-4424 Fix test with eventType provided

* PIC-4424 Fix tests by using custom EventType

* PIC-4424 Fix test

Remove eventType from the test checking the request body
as it is not provided in the request body

* PIC-4424 Refactor to use 1 S3 client

Reuse s3 client to get large hearings from a separate s3 bucket

Test using a separate local s3 bucket

* PIC-4424 Use S3 bucket from secrets

* PIC-4424 Fix test use other s3 bucket

* PIC-4424 hard code largeEventType value

* PIC-4424 Remove S3AsyncClient

* PIC-4424 Add libra event type to test

* PIC-4424 Remove unused config